# tests/integration/api/openapi.snapshot.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UW Watchlist API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/watchlist": {
      "get": {
        "tags": [
          "watchlist"
        ],
        "summary": "Get Watchlist",
        "operationId": "get_watchlist_api_watchlist_get",
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "setup",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "e.g. 'C-bull', 'C-bear', 'F-MULTI', 'NEUTRAL'",
              "title": "Setup"
            },
            "description": "e.g. 'C-bull', 'C-bear', 'F-MULTI', 'NEUTRAL'"
          },
          {
            "name": "fresh_within_minutes",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fresh Within Minutes"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchlistResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "watchlist"
        ],
        "summary": "Post Watchlist",
        "operationId": "post_watchlist_api_watchlist_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistMutation"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Post Watchlist Api Watchlist Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/watchlist/{ticker}": {
      "delete": {
        "tags": [
          "watchlist"
        ],
        "summary": "Delete Watchlist",
        "operationId": "delete_watchlist_api_watchlist__ticker__delete",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "watchlist"
        ],
        "summary": "Patch Watchlist",
        "operationId": "patch_watchlist_api_watchlist__ticker__patch",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WatchlistPatch"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Patch Watchlist Api Watchlist  Ticker  Patch"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/{ticker}": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Get Stock",
        "operationId": "get_stock_api_stock__ticker__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleStockReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/{ticker}/runs": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_stock__ticker__runs_get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response List Runs Api Stock  Ticker  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/stock/{ticker}/runs/{run_id}": {
      "get": {
        "tags": [
          "stock"
        ],
        "summary": "Get Specific Run",
        "operationId": "get_specific_run_api_stock__ticker__runs__run_id__get",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SingleStockReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/watchlist/{ticker}/rescan": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Enqueue Rescan",
        "operationId": "enqueue_rescan_api_watchlist__ticker__rescan_post",
        "parameters": [
          {
            "name": "ticker",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Ticker"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "requested_at": {
            "type": "string",
            "format": "date-time",
            "title": "Requested At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status",
          "requested_at"
        ],
        "title": "JobStatus"
      },
...
      "MarketStructure": {
        "properties": {
          "spot": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "nearest_expiry": {
            "anyOf": [
              {
                "type": "string",
                "format": "date"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nearest Expiry"
          },
          "total_call_gex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Call Gex"
          },
          "total_put_gex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Put Gex"
          },
          "net_gex": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Net Gex"
          },
          "total_call_dex_oi": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Call Dex Oi"
          },
          "total_put_dex_oi": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Put Dex Oi"
          },
          "max_pain": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pain"
          },
          "top_call_oi_strikes": {
            "items": {
              "type": "string",
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
            },
            "type": "array",
            "title": "Top Call Oi Strikes",
            "default": []
          },
          "top_put_oi_strikes": {
            "items": {
              "type": "string",
              "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
            },
            "type": "array",
            "title": "Top Put Oi Strikes",
            "default": []
          }
        },
        "type": "object",
        "title": "MarketStructure"
      },
...
      "SingleStockReport": {
        "properties": {
          "run_id": {
            "type": "integer",
            "title": "Run Id"
          },
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "short_int_note": {
            "type": "string",
            "title": "Short Int Note",
            "default": "n/a (UW endpoint does not expose %)"
          },
          "market_structure": {
            "$ref": "#/components/schemas/MarketStructure"
          },
          "volatility": {
            "$ref": "#/components/schemas/VolatilityProfile"
          },
          "flow": {
            "$ref": "#/components/schemas/FlowSnapshot"
          },
          "vrp": {
            "$ref": "#/components/schemas/VRPAssessment"
          },
          "setup": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SetupClassification"
              },
              {
                "type": "null"
              }
            ]
          },
          "trade_plan": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TradePlan"
              },
              {
                "type": "null"
              }
            ]
          },
          "dark_pool_notional": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dark Pool Notional"
          },
          "dark_pool_print_count": {
            "type": "integer",
            "title": "Dark Pool Print Count",
            "default": 0
          },
          "short_data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShortDataRow"
              },
              {
                "type": "null"
              }
            ]
          },
          "max_pain_rows": {
            "items": {
              "$ref": "#/components/schemas/MaxPainRow"
            },
            "type": "array",
            "title": "Max Pain Rows",
            "default": []
          },
          "oi_change_top": {
            "items": {
              "$ref": "#/components/schemas/OiChangeRow"
            },
            "type": "array",
            "title": "Oi Change Top",
            "default": []
          },
          "aggregates": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MarketAggregates"
              },
              {
                "type": "null"
              }
            ]
          },
          "strike_gex_curve": {
            "items": {
              "$ref": "#/components/schemas/StrikeGexBucket"
            },
            "type": "array",
            "title": "Strike Gex Curve",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "ticker",
          "generated_at",
          "market_structure",
          "volatility",
          "flow",
          "vrp"
        ],
        "title": "SingleStockReport"
      },
      "SkewBlock": {
        "properties": {
          "rr25d_30dte": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rr25D 30Dte"
          }
        },
        "type": "object",
        "title": "SkewBlock"
      },
...
      "VRPAssessment": {
        "properties": {
          "vrp": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vrp"
          },
          "signal": {
            "type": "string",
            "title": "Signal"
          },
          "note": {
            "type": "string",
            "title": "Note"
          }
        },
        "type": "object",
        "required": [
          "signal",
          "note"
        ],
        "title": "VRPAssessment"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WatchlistCard": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned"
          },
          "sort_rank": {
            "type": "integer",
            "title": "Sort Rank"
          },
          "spot": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot"
          },
          "spot_quoted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot Quoted At"
          },
          "spot_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Spot Source"
          },
          "scanned_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At"
          },
          "iv_atm": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv Atm"
          },
          "iv_rank": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Iv Rank"
          },
          "setup": {
            "$ref": "#/components/schemas/SetupBlock"
          },
          "aggression_pct": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aggression Pct"
          },
          "returns": {
            "$ref": "#/components/schemas/ReturnsBlock"
          },
          "gamma": {
            "$ref": "#/components/schemas/GammaBlock"
          },
          "skew": {
            "$ref": "#/components/schemas/SkewBlock"
          },
          "positioning": {
            "$ref": "#/components/schemas/PositioningBlock"
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "sector",
          "pinned",
          "sort_rank",
          "setup",
          "returns",
          "gamma",
          "skew",
          "positioning"
        ],
        "title": "WatchlistCard"
      },
      "WatchlistMutation": {
        "properties": {
          "ticker": {
            "type": "string",
            "title": "Ticker"
          },
          "sector": {
            "type": "string",
            "title": "Sector"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pinned": {
            "type": "boolean",
            "title": "Pinned",
            "default": false
          },
          "sort_rank": {
            "type": "integer",
            "title": "Sort Rank",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "ticker",
          "sector"
        ],
        "title": "WatchlistMutation"
      },
      "WatchlistPatch": {
        "properties": {
          "sector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sector"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "pinned": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pinned"
          },
          "sort_rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Rank"
          }
        },
        "type": "object",
        "title": "WatchlistPatch"
      },
      "WatchlistResponse": {
        "properties": {
          "scanned_at_min": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At Min"
          },
          "scanned_at_max": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At Max"
          },
          "scheduler_lag_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scheduler Lag Seconds"
          },
          "tickers": {
            "items": {
              "$ref": "#/components/schemas/WatchlistCard"
            },
            "type": "array",
            "title": "Tickers"
          }
        },
        "type": "object",
        "required": [
          "tickers"
        ],
        "title": "WatchlistResponse"
      }
    }
  }
}